requirements/base.txt
appdirs==1.4.3
argon2-cffi==16.3.0
cffi==1.10.0
coreapi==2.3.0
coreschema==0.0.4
Django==1.10.6
django-debug-toolbar==1.6
django-extensions==1.7.4
django-notifyAll==0.3
django-rest-swagger==2.1.2
djangorestframework==3.6.2
httplib2==0.10.3
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==1.0
openapi-codec==1.3.1
packaging==16.8
plivo==0.11.3
psycopg2==2.7.1
pycparser==2.17
pyparsing==2.2.0
PySocks==1.6.7
pytz==2016.10
requests==2.13.0
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.3
twilio==5.7.0
uritemplate==3.0.0
gunicorn==19.7.1
